make authorization using gate

+++ app/Http/Controllers/QuestionsController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\AskQuestionRequest;
 use App\Question;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 
 class QuestionsController extends Controller
@@ -13,6 +14,10 @@ class QuestionsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+    public function __construct()
+    {
+        $this->middleware('auth',['except'=> ['index','show']]);
+    }
     public function index()
     {
         $questions = Question::with('user')->latest()->paginate(10);
@@ -62,6 +67,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (Gate::denies('update-question',$question)){
+            abort(403);
+        }
         return view('questions.edit',compact('question'));
     }
 
@@ -87,6 +95,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (Gate::denies('delete-question',$question)){
+            abort(403);
+        }
         $question->delete();
         return  redirect()->route('questions.index')->with('success','Your question has been deleted');
     }
